+++ spanner-data-validator-java/src/main/java/com/google/migration/partitioning/PartitionRangeListFetcher.java
@@ -1,6 +1,7 @@
 package com.google.migration.partitioning;
 
 import com.google.migration.dto.PartitionRange;
+import java.math.BigDecimal;
 import java.util.List;
 import org.apache.beam.sdk.values.KV;
 
@@ -8,11 +9,11 @@ public interface PartitionRangeListFetcher {
   List<PartitionRange> getPartitionRanges(Integer partitionCount);
   public List<PartitionRange> getPartitionRangesWithCoverage(
       Integer partitionCount,
-      Integer coveragePercent);
+      BigDecimal coveragePercent);
   public List<PartitionRange> getPartitionRangesWithCoverage(String startStr,
       String endStr,
       Integer partitionCount,
-      Integer coveragePercent);
+      BigDecimal coveragePercent);
   public List<PartitionRange> getPartitionRangesWithPartitionFilter(String startStr,
       String endStr,
       Integer partitionCount,
